+++ DataProcess/Task/CClassificationTask.cpp
@@ -32,9 +32,9 @@ void CClassificationTask::generateRandomColors()
     for (size_t i=0; i<m_classNames.size(); ++i)
     {
         CColor color = {
-            (int)((double)std::rand() * factor),
-            (int)((double)std::rand() * factor),
-            (int)((double)std::rand() * factor)
+            (uchar)((double)std::rand() * factor),
+            (uchar)((double)std::rand() * factor),
+            (uchar)((double)std::rand() * factor)
         };
         m_classColors.push_back(color);
     }
